provider/aws: Add support for Skipping Final Snapshot in RDS Cluster
This works exactly the same as in DB Instance. We cannot use an empty
string to skip the final snapshot or we get the following error:

```
only alphanumeric characters and hyphens allowed in
"final_snapshot_identifier"
```

Therefore, we need to wrap this with another parameter. The we DO NOT
skip final snapshots by default as we currently create them right now
and this would change the user functionality